requirements.txt
requests>=2.28
scipy>=1.10
python-dateutil>=2.8
numpy>=1.24
